dock-manager: delay focus tracking
We track focus so that we can expand hierarchy as necessary to make
widgets visible. This is problematic when widgets are being added to
the hierarchy since we might expand things we don't intend to.

This delays it to the end of the main loop iteration so that lots of
widget additions do not harm our presentation state.