tab: start implementing custom tab without GtkToggleButton
This was getting difficult because we might want to have buttons inside
the tab. Using buttons in buttons for that is rather ugly. So instead, we
will create our own windows, handle proper margin/border/padding drawing,
and size allocations.

There is still more to do, like action state tracking which can be rather
annoying outside of gtk since we don't have the GtkActionHelper/Muxer.