tab: track prelight better
We will only do prelight when we are inside the border area of the widget.
That gives us some more space around the edge to catch a movement outside
the GDK_INPUT_OUTPUT window.

This could all be made easier with capture/bubble event propagation. We
should do that when porting to Gtk4 eventually.